Filtering Cards through 'eventType' Complete

+++ server/routes/hallMaster.js
@@ -1,29 +1,48 @@
 import { default as express } from 'express';
 const router = express.Router();
+import { ObjectId } from 'mongodb';
 
 import hallMaster from "../models/hallmaster-schema.js";
 
 router.get('/', async(req, res)=> {
 
-    const { hallCity } = req.query;
+    const { hallCity, eventId } = req.query;
+    let specificObjectId = "";
 
-    // Build the filter object based on the provided parameters
-    const filter = {};
-
-    if (hallCity) {
-        filter.hall_city = hallCity;
+    function isObjectIdFormat(str) {
+        return /^[0-9a-fA-F]{24}$/.test(str);
+      }
+    
+    if (eventId && isObjectIdFormat(eventId)) {
+        specificObjectId = new ObjectId(eventId);
+        if (!ObjectId.isValid(specificObjectId)) {
+            specificObjectId = null;
+        }
+    } else {
+        specificObjectId = null;    
     }
 
     try {
-        const hallDetails = await hallMaster.find(filter);
+        const hallDetails = await hallMaster.aggregate([
+            {
+                $match: {
+                    "hall_city": hallCity ? hallCity : {$exists: true},
+                    "hall_eventtype" : specificObjectId ? { $in: [specificObjectId] } : { $exists: true }
+                }
+            }
+        ]);
         return res.status(200).json(hallDetails); 
     } catch(error) {
         return res.status(500).json({message: error.message});
     }
-})
+});
 
 router.post('/', async (req, res) => {
     const newDocument = new hallMaster(req.body);
+
+    if(!newDocument) {
+        return res.status(404).json({message: "Request Body attachment not found!!"});
+    }
 
     try {
         const savedDocument = await newDocument.save();
@@ -37,6 +56,10 @@ router.patch('/:id', async (req, res) => {
 
     const resourceId = req.params.id;
     const updatedFields = req.body;
+
+    if(!resourceId || !updatedFields) {
+        return res.status(404).json({message: "Required fields missing!!"});
+    }
 
     try {
         const updatedResource = await hallMaster.findOneAndUpdate(
@@ -58,6 +81,10 @@ router.patch('/:id', async (req, res) => {
 router.delete("/:id", async (req, res) => {
 
     const { id } = req.params;
+
+    if(!id) {
+        return res.status(404).json({message: "Missing Id parameter!!"});
+    }
 
     try {
       const deletedHall = await hallMaster.findByIdAndDelete(id);
